document-intelligence/pyproject.toml
[project]
name = "docint"
version = "0.1.0"
description = "Barebones AI FastAPI project"
authors = [
    {name = "Raj Vasani"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "fastapi (==0.117.1)",
    "langchain (==1.0.0a6)",
    "uvicorn[standard] (>=0.36.0,<0.37.0)",
    "ollama (>=0.6.0,<0.7.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "requests >=2.32.0,<3.0.0",
    "python-multipart >=0.0.9,<0.1.0",
    "pydantic-settings >=2.4.0,<3.0.0",
    "orjson >=3.10.0,<4.0.0",
    "httpx >=0.28.0,<1.0.0",
]

[dependency-groups]
dev = [
    "ruff (==0.13.1)",
    "mypy (==1.18.1)",
    "openapi-generator-cli (>=7.15.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["src/docint_app/apis", "src/docint_app/models"] # exclude autogenerated stuff

[tool.ruff.lint]
select = [
    "E", # pycodestyle (PEP8)
    "F", # Pyflakes (static analyzer)
    "I"  # import sorting
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true